Create Introduction.md
Previously I have added an "Introduction" part in on our Research Paper to create a smooth start to the topic. I am adding that part here as an .md file as well. I am not sure if we need to use that part on our website. If we will use that part on our website this one would be a nice wallpaper (previously used by Vini as a landing page wallpaper) : static/nuclear_background.png